Reorganized internal forecasting utility functions.
